# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Manages versioned email templates and dispatches emails\nasynchronously via a DB-backed outbox to a pluggable Sender.",
        "title": "Copium Communication API",
        "contact": {
            "name": "copium maintainers"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/emails": {
            "get": {
                "description": "Returns recent outbox rows newest-first, optionally narrowed by\nstatus and a created_at window. `from`/`to` accept RFC3339 time.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "List outbox rows",
                "parameters": [
                    {
                        "type": "string",
                        "description": "queued|sending|sent|failed|dead",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 start (inclusive)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 end (exclusive)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "max rows (default 200, max 1000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/OutboxResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "bad query params",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/emails/send": {
            "post": {
                "description": "Validates params against the active template version's JSON Schema,\nresolves the recipient via custapi, renders the message, and writes\na snapshot to email_outbox. The worker performs the actual send.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "Enqueue an email",
                "parameters": [
                    {
                        "description": "Send request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SendEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/SendEmailResponse"
                        }
                    },
                    "400": {
                        "description": "invalid JSON or params failed schema",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "template or user not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "custapi unreachable",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/emails/{id}": {
            "get": {
                "description": "Returns the snapshot, status, attempts, last_error and provider\nmessage id for one queued/sent/dead email.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "Inspect an outbox row",
                "parameters": [
                    {
                        "type": "string",
                        "description": "outbox UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/OutboxResponse"
                        }
                    },
                    "400": {
                        "description": "id is not a UUID",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Liveness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Readiness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/templates": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "List templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TemplateResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Create a template",
                "parameters": [
                    {
                        "description": "Template metadata",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/TemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "code already exists",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/templates/preview": {
            "post": {
                "description": "Renders subject/body server-side using the same renderer as send.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Preview a template draft",
                "parameters": [
                    {
                        "description": "Draft to render",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PreviewTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PreviewTemplateResponse"
                        }
                    },
                    "400": {
                        "description": "invalid body, schema, or params",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/templates/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Get a template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "template UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "templates"
                ],
                "summary": "Delete a template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "template UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/templates/{id}/active-version": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Set the active version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "template UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Target version id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SetActiveVersionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "version belongs to another template",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/templates/{id}/versions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "List template versions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "template UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TemplateVersionResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Validates that params_schema is itself a compilable JSON Schema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Create a template version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "template UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Version content",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateTemplateVersionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/TemplateVersionResponse"
                        }
                    },
                    "400": {
                        "description": "invalid body or schema",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "template not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/templates/{id}/versions/{version}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Get a template version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "template UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "version number",
                        "name": "version",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TemplateVersionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CreateTemplateRequest": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
        "CreateTemplateVersionRequest": {
            "type": "object",
            "required": [
                "body_html",
                "params_schema",
                "subject"
            ],
            "properties": {
                "body_html": {
                    "type": "string"
                },
                "body_text": {
                    "type": "string"
                },
                "from_address": {
                    "type": "string"
                },
                "params_schema": {
                    "type": "object"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "not found: template \u003cuuid\u003e"
                }
            }
        },
        "OutboxResponse": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "last_error": {
                    "type": "string"
                },
                "max_attempts": {
                    "type": "integer"
                },
                "provider": {
                    "type": "string"
                },
                "provider_message_id": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "sent"
                },
                "subject": {
                    "type": "string"
                },
                "template_version_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "to_address": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
        "PreviewTemplateRequest": {
            "type": "object",
            "required": [
                "body_html",
                "params_schema",
                "subject"
            ],
            "properties": {
                "body_html": {
                    "type": "string"
                },
                "body_text": {
                    "type": "string"
                },
                "params": {
                    "type": "object"
                },
                "params_schema": {
                    "type": "object"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
        "PreviewTemplateResponse": {
            "type": "object",
            "properties": {
                "body_html": {
                    "type": "string"
                },
                "body_text": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
        "SendEmailRequest": {
            "type": "object",
            "required": [
                "template_id"
            ],
            "properties": {
                "params": {
                    "type": "object"
                },
                "template_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "to_address": {
                    "type": "string",
                    "example": "someone@example.com"
                },
                "user_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "00000000-0000-0000-0000-000000000000"
                }
            }
        },
        "SendEmailResponse": {
            "type": "object",
            "properties": {
                "outbox_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "status": {
                    "type": "string",
                    "example": "queued"
                }
            }
        },
        "SetActiveVersionRequest": {
            "type": "object",
            "required": [
                "version_id"
            ],
            "properties": {
                "version_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
        "TemplateResponse": {
            "type": "object",
            "properties": {
                "active_version_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "TemplateVersionResponse": {
            "type": "object",
            "properties": {
                "body_html": {
                    "type": "string"
                },
                "body_text": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "from_address": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "params_schema": {
                    "type": "object"
                },
                "subject": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "version": {
                    "type": "integer"
                }
            }
        }
    }
}